+++ Parser/peg_api.c
@@ -1,4 +1,4 @@
-#include "parser_interface.h"
+#include "cpython/parser_interface.h"
 
 #include "tokenizer.h"
 #include "pegen.h"
